api/requirements.txt
# =====================================================
# Alpha Laundry Management System - Python Backend
# FastAPI-based REST API with PostgreSQL
# =====================================================

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# CORS & Middleware
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0

# Development Tools (optional)
# pytest==7.4.4
# pytest-asyncio==0.23.3
# httpx==0.26.0
# black==23.12.1
# flake8==7.0.0
# mypy==1.8.0
